research-mcp/requirements.txt
mcp==1.3.0               # Anthropic MCP Python SDK
openai>=1.52.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tenacity==8.3.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.136.0
